Trait bytevec::BVSize

pub trait BVSize: Sized {
    fn from_usize(val: usize) -> Self;
    fn as_usize(&self) -> usize;
    fn max_value() -> Self;
    fn checked_add(self, rhs: Self) -> Option<Self>;
    fn get_size_of() -> Self;
}

Represents the generic integral type of the structure size indicators

Required Methods

fn from_usize(val: usize) -> Self

Returns a Self value casted from an usize value

fn as_usize(&self) -> usize

Returns an usize value casted from a Self value

fn max_value() -> Self

Returns the max value for Self

fn checked_add(self, rhs: Self) -> Option<Self>

Calls the checked_add method of self passing rhs

fn get_size_of() -> Self

Returns the returned value of std::mem::size_of for Self